Move ini-file-parsing machinery from main() to ini-file-parser
Move load_config_file() to ini-file-parser, along with some ancillary functions,
structs and related changes in headers.

This has the benefit that it will make testing this function separately more
easy.

The changes in this commit are intentionally kept to a minimum w.r.t. remains
and reorganizations, for ease of review.
