+++ avahi-daemon/main.c
@@ -40,7 +40,6 @@
 #include <time.h>
 #include <stdlib.h>
 #include <sys/time.h>
-#include <sys/resource.h>
 #include <sys/socket.h>
 
 #ifdef HAVE_INOTIFY
@@ -94,49 +93,6 @@ AvahiServer *avahi_server = NULL;
 AvahiSimplePoll *simple_poll_api = NULL;
 static char *argv0 = NULL;
 int nss_support = 0;
-
-typedef enum {
-    DAEMON_RUN,
-    DAEMON_KILL,
-    DAEMON_VERSION,
-    DAEMON_HELP,
-    DAEMON_RELOAD,
-    DAEMON_CHECK
-} DaemonCommand;
-
-typedef struct {
-    AvahiServerConfig server_config;
-    DaemonCommand command;
-    int daemonize;
-    int use_syslog;
-    char *config_file;
-#ifdef HAVE_DBUS
-    int enable_dbus;
-    int fail_on_missing_dbus;
-    unsigned n_clients_max;
-    unsigned n_objects_per_client_max;
-    unsigned n_entries_per_entry_group_max;
-#endif
-    int drop_root;
-    int set_rlimits;
-#ifdef ENABLE_CHROOT
-    int use_chroot;
-#endif
-    int modify_proc_title;
-
-    int disable_user_service_publishing;
-    int publish_resolv_conf;
-    char ** publish_dns_servers;
-    int debug;
-
-    int rlimit_as_set, rlimit_core_set, rlimit_data_set, rlimit_fsize_set, rlimit_nofile_set, rlimit_stack_set;
-    rlim_t rlimit_as, rlimit_core, rlimit_data, rlimit_fsize, rlimit_nofile, rlimit_stack;
-
-#ifdef RLIMIT_NPROC
-    int rlimit_nproc_set;
-    rlim_t rlimit_nproc;
-#endif
-} DaemonConfig;
 
 #define RESOLV_CONF "/etc/resolv.conf"
 #define BROWSE_DOMAINS_MAX 16
@@ -297,26 +253,6 @@ static void update_wide_area_servers(void) {
     avahi_server_set_wide_area_servers(avahi_server, a, n);
 }
 
-static AvahiStringList *filter_duplicate_domains(AvahiStringList *l) {
-    AvahiStringList *e, *n, *p;
-
-    if (!l)
-        return l;
-
-    for (p = l, e = l->next; e; e = n) {
-        n = e->next;
-
-        if (avahi_domain_equal((char*) e->text, (char*) l->text)) {
-            p->next = e->next;
-            avahi_free(e);
-        } else
-            p = e;
-    }
-
-    l->next = filter_duplicate_domains(l->next);
-    return l;
-}
-
 static void update_browse_domains(void) {
     AvahiStringList *l;
     int n;
@@ -336,7 +272,7 @@ static void update_browse_domains(void) {
             l = avahi_string_list_add(l, *p);
     }
 
-    l = filter_duplicate_domains(l);
+    l = avahi_ini_filter_duplicate_domains(l);
 
     avahi_server_set_browse_domains(avahi_server, l);
     avahi_string_list_free(l);
@@ -546,448 +482,6 @@ static int parse_command_line(DaemonConfig *c, int argc, char *argv[]) {
     }
 
     return 0;
-}
-
-static int is_yes(const char *s) {
-    assert(s);
-
-    return *s == 'y' || *s == 'Y' || *s == '1' || *s == 't' || *s == 'T';
-}
-
-static int parse_unsigned(const char *s, unsigned *u) {
-    char *e = NULL;
-    unsigned long ul;
-    unsigned k;
-
-    errno = 0;
-    ul = strtoul(s, &e, 0);
-
-    if (!e || *e || errno != 0)
-        return -1;
-
-    k = (unsigned) ul;
-
-    if ((unsigned long) k != ul)
-        return -1;
-
-    *u = k;
-    return 0;
-}
-
-static int parse_usec(const char *s, AvahiUsec *u) {
-    char *e = NULL;
-    unsigned long long ull;
-    AvahiUsec k;
-
-    errno = 0;
-    ull = strtoull(s, &e, 0);
-
-    if (!e || *e || errno != 0)
-        return -1;
-
-    k = (AvahiUsec) ull;
-
-    if ((unsigned long long) k != ull)
-        return -1;
-
-    *u = k;
-    return 0;
-}
-
-static char *get_machine_id(void) {
-    int fd;
-    char buf[32];
-
-    fd = open("/etc/machine-id", O_RDONLY|O_CLOEXEC|O_NOCTTY);
-    if (fd == -1 && errno == ENOENT)
-        fd = open("/var/lib/dbus/machine-id", O_RDONLY|O_CLOEXEC|O_NOCTTY);
-    if (fd == -1)
-        return NULL;
-
-    /* File is on a filesystem so we never get EINTR or partial reads */
-    if (read(fd, buf, sizeof buf) != sizeof buf) {
-        close(fd);
-        return NULL;
-    }
-    close(fd);
-
-    /* Contents can be lower, upper and even mixed case so normalize */
-    avahi_strdown(buf);
-
-    return avahi_strndup(buf, sizeof buf);
-}
-
-static int load_config_file(DaemonConfig *c, const char* config_file) {
-    int r = -1;
-    AvahiIniFile *f;
-    AvahiIniFileGroup *g;
-
-    assert(c);
-    assert(config_file);
-
-    if (!(f = avahi_ini_file_load(config_file)))
-        goto finish;
-
-    for (g = f->groups; g; g = g->groups_next) {
-
-        if (strcasecmp(g->name, "server") == 0) {
-            AvahiIniFilePair *p;
-
-            for (p = g->pairs; p; p = p->pairs_next) {
-
-                if (strcasecmp(p->key, "host-name") == 0) {
-                    avahi_free(c->server_config.host_name);
-                    c->server_config.host_name = avahi_strdup(p->value);
-                } else if (strcasecmp(p->key, "domain-name") == 0) {
-                    avahi_free(c->server_config.domain_name);
-                    c->server_config.domain_name = avahi_strdup(p->value);
-                } else if (strcasecmp(p->key, "browse-domains") == 0) {
-                    char **e, **t;
-                    char *line;
-
-                    avahi_log_debug("browse-domains: processing line: '%s'", p->value);
-                    e = avahi_split_csv(p->value);
-                    if (!e || (e[0] && strcmp(e[0], "") == 0)) {
-                        /* reset if the string is null or empty */
-                        avahi_log_debug("browse-domains reset");
-
-                        avahi_string_list_free(c->server_config.browse_domains);
-                        c->server_config.browse_domains = NULL;
-                    } else {
-                        for (t = e; *t; t++) {
-                            char cleaned[AVAHI_DOMAIN_NAME_MAX];
-
-                            if (!avahi_normalize_name(*t, cleaned, sizeof(cleaned))) {
-                                avahi_log_error("Invalid domain name \"%s\" for key \"%s\" in group \"%s\"\n", *t, p->key, g->name);
-                                avahi_strfreev(e);
-                                goto finish;
-                            }
-
-                            avahi_log_debug("browse-domains: adding '%s'", cleaned);
-                            c->server_config.browse_domains = avahi_string_list_add(c->server_config.browse_domains, cleaned);
-                        }
-                    }
-                    avahi_strfreev(e);
-
-                    c->server_config.browse_domains = filter_duplicate_domains(c->server_config.browse_domains);
-
-                    line = avahi_string_list_to_string(c->server_config.browse_domains);
-                    avahi_log_debug("browse-domains: after processing complete line: '%s'", line);
-                    avahi_free(line);
-
-                } else if (strcasecmp(p->key, "use-ipv4") == 0)
-                    c->server_config.use_ipv4 = is_yes(p->value);
-                else if (strcasecmp(p->key, "use-ipv6") == 0)
-                    c->server_config.use_ipv6 = is_yes(p->value);
-                else if (strcasecmp(p->key, "check-response-ttl") == 0)
-                    c->server_config.check_response_ttl = is_yes(p->value);
-                else if (strcasecmp(p->key, "allow-point-to-point") == 0)
-                    c->server_config.allow_point_to_point = is_yes(p->value);
-                else if (strcasecmp(p->key, "use-iff-running") == 0)
-                    c->server_config.use_iff_running = is_yes(p->value);
-                else if (strcasecmp(p->key, "disallow-other-stacks") == 0)
-                    c->server_config.disallow_other_stacks = is_yes(p->value);
-                else if (strcasecmp(p->key, "host-name-from-machine-id") == 0) {
-                    if (*(p->value) == 'y' || *(p->value) == 'Y') {
-                        char *machine_id = get_machine_id();
-                        if (machine_id != NULL) {
-                            avahi_free(c->server_config.host_name);
-                            c->server_config.host_name = machine_id;
-                        }
-                    }
-                }
-#ifdef HAVE_DBUS
-                else if (strcasecmp(p->key, "enable-dbus") == 0) {
-
-                    if (*(p->value) == 'w' || *(p->value) == 'W') {
-                        c->fail_on_missing_dbus = 0;
-                        c->enable_dbus = 1;
-                    } else if (*(p->value) == 'y' || *(p->value) == 'Y') {
-                        c->fail_on_missing_dbus = 1;
-                        c->enable_dbus = 1;
-                    } else {
-                        c->enable_dbus = 0;
-                    }
-                }
-#endif
-                else if (strcasecmp(p->key, "allow-interfaces") == 0) {
-                    char **e, **t;
-                    char *line;
-
-                    avahi_log_debug("allow-interfaces: processing line: '%s'", p->value);
-                    e = avahi_split_csv(p->value);
-                    if (!e || (e[0] && strcmp(e[0], "") == 0)) {
-                        /* reset if the string is null or empty */
-                        line = avahi_string_list_to_string(c->server_config.allow_interfaces);
-                        avahi_log_debug("allow-interfaces reset, content was: '%s'", line);
-                        avahi_free(line);
-
-                        avahi_string_list_free(c->server_config.allow_interfaces);
-                        c->server_config.allow_interfaces = NULL;
-                    } else {
-                        for (t = e; *t; t++) {
-                            if (t[0] && strcmp(t[0], "") != 0) {
-                                avahi_log_debug("allow-interfaces: adding '%s'", t[0]);
-                                c->server_config.allow_interfaces = avahi_string_list_add(c->server_config.allow_interfaces, *t);
-                            } else
-                                avahi_log_debug("allow-interfaces: skipping empty value '%s'", t[0]);
-                        }
-                    }
-                    avahi_strfreev(e);
-
-                    line = avahi_string_list_to_string(c->server_config.allow_interfaces);
-                    avahi_log_debug("allow-interfaces: after processing complete line: '%s'", line);
-                    avahi_free(line);
-
-                } else if (strcasecmp(p->key, "deny-interfaces") == 0) {
-                    char **e, **t;
-                    char *line;
-
-                    avahi_log_debug("deny-interfaces: processing line: '%s'", p->value);
-                    e = avahi_split_csv(p->value);
-                    if (!e || (e[0] && strcmp(e[0], "") == 0)) {
-                        /* reset if the string is null or empty */
-                        line = avahi_string_list_to_string(c->server_config.deny_interfaces);
-                        avahi_log_debug("deny-interfaces reset, content was: '%s'", line);
-                        avahi_free(line);
-
-                        avahi_string_list_free(c->server_config.deny_interfaces);
-                        c->server_config.deny_interfaces = NULL;
-                    } else {
-                        for (t = e; *t; t++)
-                            if (t[0] && strcmp(t[0], "") != 0) {
-                                avahi_log_debug("deny-interfaces: adding '%s'", t[0]);
-                                c->server_config.deny_interfaces = avahi_string_list_add(c->server_config.deny_interfaces, *t);
-                            } else
-                                avahi_log_debug("deny-interfaces: skipping empty value '%s'", t[0]);
-                    }
-                    avahi_strfreev(e);
-
-                    line = avahi_string_list_to_string(c->server_config.deny_interfaces);
-                    avahi_log_debug("deny-interfaces: after processing complete line: '%s'", line);
-                    avahi_free(line);
-
-                } else if (strcasecmp(p->key, "ratelimit-interval-usec") == 0) {
-                    AvahiUsec k;
-
-                    if (parse_usec(p->value, &k) < 0) {
-                        avahi_log_error("Invalid ratelimit-interval-usec setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->server_config.ratelimit_interval = k;
-
-                } else if (strcasecmp(p->key, "ratelimit-burst") == 0) {
-                    unsigned k;
-
-                    if (parse_unsigned(p->value, &k) < 0) {
-                        avahi_log_error("Invalid ratelimit-burst setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->server_config.ratelimit_burst = k;
-
-                } else if (strcasecmp(p->key, "cache-entries-max") == 0) {
-                    unsigned k;
-
-                    if (parse_unsigned(p->value, &k) < 0) {
-                        avahi_log_error("Invalid cache-entries-max setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->server_config.n_cache_entries_max = k;
-#ifdef HAVE_DBUS
-                } else if (strcasecmp(p->key, "clients-max") == 0) {
-                    unsigned k;
-
-                    if (parse_unsigned(p->value, &k) < 0) {
-                        avahi_log_error("Invalid clients-max setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->n_clients_max = k;
-                } else if (strcasecmp(p->key, "objects-per-client-max") == 0) {
-                    unsigned k;
-
-                    if (parse_unsigned(p->value, &k) < 0) {
-                        avahi_log_error("Invalid objects-per-client-max setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->n_objects_per_client_max = k;
-                } else if (strcasecmp(p->key, "entries-per-entry-group-max") == 0) {
-                    unsigned k;
-
-                    if (parse_unsigned(p->value, &k) < 0) {
-                        avahi_log_error("Invalid entries-per-entry-group-max setting %s", p->value);
-                        goto finish;
-                    }
-
-                    c->n_entries_per_entry_group_max = k;
-#endif
-                } else {
-                    avahi_log_error("Invalid configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-                    goto finish;
-                }
-            }
-
-        } else if (strcasecmp(g->name, "publish") == 0) {
-            AvahiIniFilePair *p;
-
-            for (p = g->pairs; p; p = p->pairs_next) {
-
-                if (strcasecmp(p->key, "publish-addresses") == 0)
-                    c->server_config.publish_addresses = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-hinfo") == 0)
-                    c->server_config.publish_hinfo = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-workstation") == 0)
-                    c->server_config.publish_workstation = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-domain") == 0)
-                    c->server_config.publish_domain = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-resolv-conf-dns-servers") == 0)
-                    c->publish_resolv_conf = is_yes(p->value);
-                else if (strcasecmp(p->key, "disable-publishing") == 0)
-                    c->server_config.disable_publishing = is_yes(p->value);
-                else if (strcasecmp(p->key, "disable-user-service-publishing") == 0)
-                    c->disable_user_service_publishing = is_yes(p->value);
-                else if (strcasecmp(p->key, "add-service-cookie") == 0)
-                    c->server_config.add_service_cookie = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-dns-servers") == 0) {
-                    char **e, **t;
-
-                    avahi_log_debug("publish-dns-servers: processing line: '%s'", p->value);
-                    e = avahi_split_csv(p->value);
-                    if (!e || (e[0] && strcmp(e[0], "") == 0)) {
-                        /* reset if the string is null or empty */
-                        avahi_log_debug("publish-dns-servers reset");
-
-                        avahi_strfreev(c->publish_dns_servers);
-                        c->publish_dns_servers = NULL;
-                    } else {
-                        c->publish_dns_servers = e;
-                        for (t = e; *t; t++) {
-                            if (t[0] && strcmp(t[0], "") != 0)
-                                avahi_log_debug("publish_dns_servers: adding '%s'", t[0]);
-                            else
-                                avahi_log_debug("publish_dns_servers: skipping empty value '%s'", t[0]);
-                        }
-                    }
-                } else if (strcasecmp(p->key, "publish-a-on-ipv6") == 0)
-                    c->server_config.publish_a_on_ipv6 = is_yes(p->value);
-                else if (strcasecmp(p->key, "publish-aaaa-on-ipv4") == 0)
-                    c->server_config.publish_aaaa_on_ipv4 = is_yes(p->value);
-                else {
-                    avahi_log_error("Invalid configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-                    goto finish;
-                }
-            }
-
-        } else if (strcasecmp(g->name, "wide-area") == 0) {
-            AvahiIniFilePair *p;
-
-            for (p = g->pairs; p; p = p->pairs_next) {
-
-                if (strcasecmp(p->key, "enable-wide-area") == 0)
-                    c->server_config.enable_wide_area = is_yes(p->value);
-                else {
-                    avahi_log_error("Invalid configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-                    goto finish;
-                }
-            }
-
-        } else if (strcasecmp(g->name, "reflector") == 0) {
-            AvahiIniFilePair *p;
-
-            for (p = g->pairs; p; p = p->pairs_next) {
-
-                if (strcasecmp(p->key, "enable-reflector") == 0)
-                    c->server_config.enable_reflector = is_yes(p->value);
-                else if (strcasecmp(p->key, "reflect-ipv") == 0)
-                    c->server_config.reflect_ipv = is_yes(p->value);
-                else if (strcasecmp(p->key, "reflect-filters") == 0) {
-                    char **e, **t;
-                    char *line;
-
-                    avahi_log_debug("reflect-filters: processing line: '%s'", p->value);
-                    e = avahi_split_csv(p->value);
-                    if (!e || (e[0] && strcmp(e[0], "") == 0)) {
-                        /* reset if the string is null or empty */
-                        avahi_log_debug("reflect-filters reset");
-
-                        avahi_string_list_free(c->server_config.reflect_filters);
-                        c->server_config.reflect_filters = NULL;
-                    } else {
-                        for (t = e; *t; t++) {
-                            if (t[0] && strcmp(t[0], "") != 0) {
-                                avahi_log_debug("reflect-filters: adding '%s'", t[0]);
-                                c->server_config.reflect_filters = avahi_string_list_add(c->server_config.reflect_filters, *t);
-                            } else
-                                avahi_log_debug("reflect-filters: skipping empty value '%s'", t[0]);
-                        }
-                    }
-                    avahi_strfreev(e);
-
-                    line = avahi_string_list_to_string(c->server_config.reflect_filters);
-                    avahi_log_debug("reflect-filters: after processing complete line: '%s'", line);
-                    avahi_free(line);
-                }
-                else {
-                    avahi_log_error("Invalid configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-                    goto finish;
-                }
-            }
-
-        } else if (strcasecmp(g->name, "rlimits") == 0) {
-            AvahiIniFilePair *p;
-
-            for (p = g->pairs; p; p = p->pairs_next) {
-
-                if (strcasecmp(p->key, "rlimit-as") == 0) {
-                    c->rlimit_as_set = 1;
-                    c->rlimit_as = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-core") == 0) {
-                    c->rlimit_core_set = 1;
-                    c->rlimit_core = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-data") == 0) {
-                    c->rlimit_data_set = 1;
-                    c->rlimit_data = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-fsize") == 0) {
-                    c->rlimit_fsize_set = 1;
-                    c->rlimit_fsize = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-nofile") == 0) {
-                    c->rlimit_nofile_set = 1;
-                    c->rlimit_nofile = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-stack") == 0) {
-                    c->rlimit_stack_set = 1;
-                    c->rlimit_stack = atoi(p->value);
-                } else if (strcasecmp(p->key, "rlimit-nproc") == 0) {
-#ifdef RLIMIT_NPROC
-                    c->rlimit_nproc_set = 1;
-                    c->rlimit_nproc = atoi(p->value);
-#else
-                    avahi_log_error("Ignoring configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-#endif
-                } else {
-                    avahi_log_error("Invalid configuration key \"%s\" in group \"%s\"\n", p->key, g->name);
-                    goto finish;
-                }
-
-            }
-
-        } else {
-            avahi_log_error("Invalid configuration file group \"%s\".\n", g->name);
-            goto finish;
-        }
-    }
-
-    r = 0;
-
-finish:
-
-    if (f)
-        avahi_ini_file_free(f);
-
-    return r;
 }
 
 static void log_function(AvahiLogLevel level, const char *txt) {
@@ -1704,7 +1198,7 @@ int main(int argc, char *argv[]) {
 
         main_config_file = config.config_file ? config.config_file : AVAHI_CONFIG_FILE;
         avahi_log_debug("Loading main conf: '%s'", main_config_file);
-        if (load_config_file(&config, main_config_file) < 0)
+        if (avahi_ini_file_parse(&config, main_config_file) < 0)
             goto finish;
 
         snprintf_count = snprintf(confd_path, sizeof(confd_path), "%s.d", main_config_file);
@@ -1725,7 +1219,7 @@ int main(int argc, char *argv[]) {
                     goto finish;
                 }
                 avahi_log_debug("- %s", confd_file);
-                if (load_config_file(&config, confd_file) < 0) {
+                if (avahi_ini_file_parse(&config, confd_file) < 0) {
                     avahi_log_error("Could not load conf.d file: '%s'", confd_file);
                     avahi_strfreev(confd_files);
                     goto finish;
